Removed **provisional** from computed_properties (#39543)
* removed **provisional** from computed_properties

* updated changelog

* added computed_properties keyword argument to function def

* annotated computed_properties tests

* added replace computed_properties feature with tests

* added replace computed_properties feature with tests

* changed test class annotation to cosmosQuery, removed extra annotations

---------
