+++ sdk/cosmos/azure-cosmos/CHANGELOG.md
@@ -3,12 +3,14 @@
 ### 4.9.1b5 (Unreleased)
 
 #### Features Added
+* Added ability to replace `computed_properties` through `replace_container` method. See [PR 39543](https://github.com/Azure/azure-sdk-for-python/pull/39543)
 
 #### Breaking Changes
 
 #### Bugs Fixed
 
 #### Other Changes
+* Un-marked `computed_properties` keyword as **provisional**. See [PR 39543](https://github.com/Azure/azure-sdk-for-python/pull/39543)
 
 ### 4.9.1b4 (2025-02-06)
