+++ sdk/cosmos/azure-cosmos/test/test_query_async.py
@@ -489,76 +489,6 @@ class TestQueryAsync(unittest.IsolatedAsyncioTestCase):
         # verify a second time
         assert len(token.encode('utf-8')) <= 1024
 
-    @pytest.mark.cosmosQuery
-    @pytest.mark.skip
-    async def test_computed_properties_query(self):
-        computed_properties = [{'name': "cp_lower", 'query': "SELECT VALUE LOWER(c.db_group) FROM c"},
-                               {'name': "cp_power",
-                                'query': "SELECT VALUE POWER(c.val, 2) FROM c"},
-                               {'name': "cp_str_len", 'query': "SELECT VALUE LENGTH(c.stringProperty) FROM c"}]
-        items = [
-            {'id': str(uuid.uuid4()), 'pk': 'test', 'val': 5, 'stringProperty': 'prefixOne', 'db_group': 'GroUp1'},
-            {'id': str(uuid.uuid4()), 'pk': 'test', 'val': 5, 'stringProperty': 'prefixTwo', 'db_group': 'GrOUp1'},
-            {'id': str(uuid.uuid4()), 'pk': 'test', 'val': 5, 'stringProperty': 'randomWord1', 'db_group': 'GroUp2'},
-            {'id': str(uuid.uuid4()), 'pk': 'test', 'val': 5, 'stringProperty': 'randomWord2', 'db_group': 'groUp1'},
-            {'id': str(uuid.uuid4()), 'pk': 'test', 'val': 5, 'stringProperty': 'randomWord3', 'db_group': 'GroUp3'},
-            {'id': str(uuid.uuid4()), 'pk': 'test', 'val': 5, 'stringProperty': 'randomWord4', 'db_group': 'GrOUP1'},
-            {'id': str(uuid.uuid4()), 'pk': 'test', 'val': 5, 'stringProperty': 'randomWord5', 'db_group': 'GroUp2'},
-            {'id': str(uuid.uuid4()), 'pk': 'test', 'val': 0, 'stringProperty': 'randomWord6', 'db_group': 'group1'},
-            {'id': str(uuid.uuid4()), 'pk': 'test', 'val': 3, 'stringProperty': 'randomWord7', 'db_group': 'group2'},
-            {'id': str(uuid.uuid4()), 'pk': 'test', 'val': 2, 'stringProperty': 'randomWord8', 'db_group': 'GroUp3'}
-        ]
-        created_collection = await self.created_db.create_container(
-            "computed_properties_query_test_" + str(uuid.uuid4()),
-            PartitionKey(path="/pk"),
-            computed_properties=computed_properties)
-
-        # Create Items
-        for item in items:
-            await created_collection.create_item(body=item)
-
-        # Check if computed properties were set
-        container_properties = await created_collection._get_properties()
-        assert computed_properties == container_properties["computedProperties"]
-
-        # Test 0: Negative test, test if using non-existent computed property
-        queried_items = [q async for q in
-                         created_collection.query_items(query='Select * from c Where c.cp_upper = "GROUP2"',
-                                                        partition_key="test")]
-        assert len(queried_items) == 0
-
-        # Test 1: Test first computed property
-        queried_items = [q async for q in
-                         created_collection.query_items(query='Select * from c Where c.cp_lower = "group1"',
-                                                        partition_key="test")]
-        assert len(queried_items) == 5
-
-        # Test 1 Negative: Test if using non-existent string in group property returns nothing
-        queried_items = [q async for q in
-                         created_collection.query_items(query='Select * from c Where c.cp_lower = "group4"',
-                                                        partition_key="test")]
-        assert len(queried_items) == 0
-
-        # Test 2: Test second computed property
-        queried_items = [q async for q in created_collection.query_items(query='Select * from c Where c.cp_power = 25',
-                                                                         partition_key="test")]
-        assert len(queried_items) == 7
-
-        # Test 2 Negative: Test Non-Existent POWER
-        queried_items = [q async for q in created_collection.query_items(query='Select * from c Where c.cp_power = 16',
-                                                                         partition_key="test")]
-        assert len(queried_items) == 0
-
-        # Test 3: Test Third Computed Property
-        queried_items = [q async for q in created_collection.query_items(query='Select * from c Where c.cp_str_len = 9',
-                                                                         partition_key="test")]
-        assert len(queried_items) == 2
-
-        # Test 3 Negative: Test Str length that isn't there
-        queried_items = [q async for q in created_collection.query_items(query='Select * from c Where c.cp_str_len = 3',
-                                                                         partition_key="test")]
-        assert len(queried_items) == 0
-
     async def test_cosmos_query_retryable_error_async(self):
         async def query_items(database):
             # Tests to make sure 429 exception is surfaced when retries run out in the first page of a query.
